+++ src/plugin/engine/engine.hpp
@@ -60,6 +60,8 @@ struct Engine {
 
     package::InstallJobs package_install_jobs {};
 
+    DynamicArray<char> attribution_text {Malloc::Instance()}; // empty if none needed
+
     // IMPORTANT: debug-only, remove this
     DynamicArrayBounded<char, 200> state_change_description {};